Change Summary

This function is designed to be run on a data frame after other functions from the NutritionTools package have been run on it. This function scans the comments column for the preset comments left by other functions from the package, and produces a summary of them, and how often they appear.

Usage

Change_Summary(df, comment_col = "comments", detailed = FALSE)

Arguments

df

Required - the data.frame which contains the food items that have been changed by functions from NutritionTools.

comment_col

Required - default: 'comments' - The name of the column within df that contains comments.

detailed

Required - default: FALSE - TRUE or FALSE. If set to TRUE, then the Change Summary will examine the exact comments as they appear in the data frame. If set to FALSE, as it is by default, comments which have unique identifiers (values, imputation IDs etc) have those aspects removed, to reduce the number of return types.

Value

A console output detailing how many times one of the comments generated by functions in NutritionTools appears in the df.