ENERCKj calculator A function used to calculate Energy values in kj

This function has two modes. Either it works as a basic calculator - The values for Total Protein in grams, Total Fat in grams (ideally combined), Available Carbohydrate in grams, Fibre, Total Dietary in grams and Alcohol in grams are combined to find the Energy in kj.

Alternatively, this function works as a data.frame calculator. If the 'df' input is used, then the Nutrient inputs should be the names of the columns in 'df' which contains the details of their namesakes. The result of this function will be a new column in df with the ENERCKj_calculated values.

Usage

ENERCKj_calculator(
  Protein = "PROCNTg",
  Fat = "FAT_g_combined",
  Carbohydrate = "CHOAVLg",
  Fibre = "FIBTGg_combined",
  Alcohol = "ALCg",
  df,
  comment_col = "comments",
  Stop_If_Missing = TRUE,
  Assume_Zero_Alcohol = FALSE,
  Show_NA_outputs = TRUE
)

Arguments

Protein

Required - default: "PROTg" - If using 'df', then this should be the name of the column that contains Protein values in grams. If NOT using 'df', this should be the number or list of numbers representing Protein values you are looking to calculate an Energy value using.

Fat

Required - default: "FATg_combined" - If using 'df', then this should be the name of the column that contains Total Fat values in grams. If NOT using 'df', this should be the number or list of numbers representing Fat values you are looking to calculate an Energy value using.

Carbohydrate

Required - default: "CHOAVLg" - If using 'df', then this should be the name of the column that contains Available Carbohydrate values in grams. If NOT using 'df', this should be the number or list of numbers representing Available Carbohydrate values you are looking to calculate an Energy value using.

Fibre

Required - default: "FIBTGg" - If using 'df', then this should be the name of the column that contains Fibre values in grams. If NOT using 'df', this should be the number or list of numbers representing Fibre values you are looking to calculate an Energy value using.

Alcohol

Required - default: "ALCg" - If using 'df', then this should be the name of the column that contains Alcohol values in grams. If NOT using 'df', this should be the number or list of numbers representing Alcohol values you are looking to calculate an Energy value using. Note: This is the only Optional input. All items in all other nutrient parameters must have a valid value for the calculation - however, if Stop_If_Missing is set to FALSE (as it is by default), then ALcohol values can contain blanks, which will be reset to 0.

df

Optional - If using on a data.frame, then this input should be the data.frame in question. If left blank, then the standard calculation will take place instead.

comment_col

Optional - default: "comments" - only relevant if 'df' is in use. The name of the comment column in the df. If there isn't one present, then the function will create one to record the calculation details.

Assume_Zero_Alcohol

Required - default: FALSE - Either TRUE or FALSE. If set to TRUE, and 'Stop_If_Missing' is set to FALSE, then missing alcohol values will be set to 0, and a comment added if running the data.frame calculation. if set to FALSE, then the Alcohol value will not change.

Show_NA_Outputs

Required - default: TRUE - Either TRUE or FALSE. If set to TRUE then if any NA values are generated for ENERCKj_calculated, a data.frame of these values will be shown.

Value

If using the data.frame input, the return will be a data.frame with a newly generated ENERCKj_calculated column. If not using the data.frame input, then the return will be ENERCKj_calculated values calculated from the input values put into the function.

Examples