Sum of Proximate Calculator

Calculates SOPg_calculated = (WATERg + PROCNTg + FAT_g_standardised + CHOAVLDFg_standardised + FIBTGg_standardised_column + ALCg +ASHg). Column names are case sensitive and an error is returned if not found.

Usage

SOPg_calculator(
  df,
  WATERg_column = "WATERg",
  PROCNTg_column = "PROCNTg",
  FAT_g_standardised_column = "FAT_g_standardised",
  CHOAVLDFg_standardised_column = "CHOAVLDFg_standardised",
  FIBTGg_standardised_column = "FIBTGg_standardised",
  ALCg_column = "ALCg",
  ASHg_column = "ASHg",
  comment = T,
  comment_col = "comments",
  OutsideBoundsReplacement = "none",
  LowerBound = 93,
  UpperBound = 107,
  OutsideBoundsDF = F
)

Arguments

df

Required - the data.frame the data is currently stored in.

WATERg_column

Required - default: 'WATERg' - The name of the column containing Water/moisture content in grams per 100g of Edible Portion (EP).

PROCNTg_column

Required - default: 'PROCNTg' - Protein in grams per 100g of Edible Portion (EP), as reported in the original FCT and assumed to be calculated from nitrogen (NTg) content.

FAT_g_standardised_column

Required - default: 'FAT_g_standardised' - Fat content, unknown method of calculation, in grams per 100g of Edible Portion (EP).

CHOAVLDFg_standardised_column

Required - default: 'CHOAVLDFg_standardised' - Available carbohydrates calculated by difference, in grams per 100g of Edible Portion (EP).

FIBTGg_standardised_column

Required - default: 'FIBTGg_standardised' - Fibre content from combined Tagnames, with preference of Total dietary fibre by AOAC Prosky method, expressed in grams per 100g of Edible Portion (EP).

ALCg_column

Required - default: 'ALCg' - Alcohol in grams per 100g of Edible Portion (EP).

ASHg_column

Required - default: 'ASHg' - Ashes in grams per 100g of Edible Portion (EP).

comment

Optional - default: T - TRUE or FALSE. If comment is set to TRUE (as it is by default), when the function is run a comment describing the source of the SOPg_calculated column is added to the comment_col. If no comment_col is selected, and comment T, one is created, called comments.

comment_col

Optional - default: 'comments' - A potential input variable; the column which contains the metadata comments for the food item in question. Not required if the comment parameter is set to FALSE.

OutsideBoundsReplacement

Optional - default: 'none' - Options are 'round', NA, 'remove', or 'none'. Choose what happens to values that are outside of the bounds. The ranges are set to FAO standards: 93-107 is considered acceptable. This parameter decides what happens to those values less than 93, or over 107. If set to round, then outside of bound values are set to the closest acceptable value (e.g. 90 -> 93, 111 -> 107. If set to NA, they are replaced with NA. if set to 'remove', then those rows are removed (including NA results). if set to 'none', then they are left as the out of bound values.

LowerBound

Optional - default: 93 - Integer value. Sets the lower boundary for acceptable SOPg_calculated values, and therefore determines the values affected by OutsideBoundsReplacement and OutsideBoundsDF. FAO standards list 93 as the lower boundary for acceptable values, and 95 as the lower boundary for preferred values.

UpperBound

Optional - default: 107 - Integer value. Sets the upper boundary for acceptable SOPg_calculated values, and therefore determines the values affected by OutsideBoundsReplacement and OutsideBoundsDF. FAO standards list 107 as the upper boundary for acceptable values, and 105 as the upper boundary for preferred values.

OutsideBoundsDF

Optional - default: F - TRUE or FALSE. If set to TRUE, Then the output switches from being a copy of the input df with the the SOPg_calculated column to a subset of that dataframe only showing SOPg_calculated values that are out of bounds, for manual inspection.

Value

Original FCT dataset with a new SOPg_calculated column.

Examples

# Two example data.frames have been prepared to illustrate the
# SOPg_calculator. The first is a dataset of fictional food values to
# illustrate the various options in the function, and the second is a dataset
# with non-standard column names, to show how to specify columns.

# This is the first data.frame - before the SOPg_calculator has been used on it.
View(SOP_example_df)
#> Warning: unable to open display
#> Error in .External2(C_dataviewer, x, title): unable to start data viewer

# First, an example of the standard usecase - calculate the SOPg_calculated
# value, without modifying out of bounds values.
nothing_results <- SOPg_calculator(SOP_example_df, OutsideBoundsReplacement = "none")
#> Error in SOPg_calculator(SOP_example_df, OutsideBoundsReplacement = "none"): could not find function "SOPg_calculator"
View(nothing_results)
#> Error in eval(expr, envir, enclos): object 'nothing_results' not found
# See the changes - the addition of the SOPg_calculated column, and the
# additions to the comments column.
#
#
# The second example shows the results when the Replacement option is set to NA
NA_results <- SOPg_calculator(SOP_example_df, OutsideBoundsReplacement = NA)
#> Error in SOPg_calculator(SOP_example_df, OutsideBoundsReplacement = NA): could not find function "SOPg_calculator"
View(NA_results)
#> Error in eval(expr, envir, enclos): object 'NA_results' not found
# Check the SOP column and comments column again - see how values outside of
# bounds have been replaced with NA, and a note of this change logged in the
# comments column.
#
#
# The third example shows the results when the Replacement option is set to 'remove'
remove_results <- SOPg_calculator(SOP_example_df, OutsideBoundsReplacement = "remove")
#> Error in SOPg_calculator(SOP_example_df, OutsideBoundsReplacement = "remove"): could not find function "SOPg_calculator"
View(remove_results)
#> Error in eval(expr, envir, enclos): object 'remove_results' not found
# See how the out of bounds values have been removed.
#
#
# The fourth example is of the rounding results.
rounding_results <- SOPg_calculator(SOP_example_df, OutsideBoundsReplacement = "round")
#> Error in SOPg_calculator(SOP_example_df, OutsideBoundsReplacement = "round"): could not find function "SOPg_calculator"
View(rounding_results)
#> Error in eval(expr, envir, enclos): object 'rounding_results' not found
# Look at the SOP_standardised values - and see how they've been capped to the bounds
# if they would have been out fo bounds, with a note of the change in the comments.
#
#
# The fifth example is of the out of bounds dataframe - an option useful for identifying
# and examining out of bounds results.
OoB_DF_results <- SOPg_calculator(SOP_example_df, OutsideBoundsDF = T)
#> Error in SOPg_calculator(SOP_example_df, OutsideBoundsDF = T): could not find function "SOPg_calculator"
View(OoB_DF_results)
#> Error in eval(expr, envir, enclos): object 'OoB_DF_results' not found
# Only the out of bounds results are present, in their original form, for inspection.
#
#
# The sixth example is of the SOPg_calculator working on a dataframe with non-standard
# column names. It uses a modified example data frame, shown below.
View(SOP_example_df_nonstandard)
#> Warning: unable to open display
#> Error in .External2(C_dataviewer, x, title): unable to start data viewer
# Notice how the column names are different, and differ from the assumed names.
#
#
# Because of the different names, the column names for each input must be specified.
nothing_results_NonStandardInput <- SOPg_calculator(
SOP_example_df_nonstandard,
WATERg_column = "Water_values_g",
PROCNTg_column = "PROCNT_values_g",
FAT_g_standardised_column = "FAT_values_g_standardised",
CHOAVLDFg_standardised_column = "CHOAVLDF_values_g_standardised",
FIBTGg_standardised_column = "FIBTG_values_g_standardised",
ALCg_column = "ALC_values_g",
ASHg_column = "ASH_values_g",
comment_col = "comments_column",
LowerBound = 97,
UpperBound = 103,
OutsideBoundsReplacement = "nothing")
#> Error in SOPg_calculator(SOP_example_df_nonstandard, WATERg_column = "Water_values_g",     PROCNTg_column = "PROCNT_values_g", FAT_g_standardised_column = "FAT_values_g_standardised",     CHOAVLDFg_standardised_column = "CHOAVLDF_values_g_standardised",     FIBTGg_standardised_column = "FIBTG_values_g_standardised",     ALCg_column = "ALC_values_g", ASHg_column = "ASH_values_g",     comment_col = "comments_column", LowerBound = 97, UpperBound = 103,     OutsideBoundsReplacement = "nothing"): could not find function "SOPg_calculator"
View(nothing_results_NonStandardInput)
#> Error in eval(expr, envir, enclos): object 'nothing_results_NonStandardInput' not found
# The resulting SOPg_calculated column is the same as in the first example, despite the
# different names - although, due to the shift in the bounds, the warning message is not.