Imaginary Food Items

An imaginary dataset of Food Items used to demonstrate various functions

Usage

breakfast_df

Format

A data frame with 10 rows and 21 columns:

food_code

Imaginary Food ID

food_name

Food names to attach the imaginary nutritional data to

WATERg

Imaginary Water values

PROCNTg

Imaginary Protein values

FAT_g_combined

Imaginary Fat values

CHOAVLg

Imaginary Carbohyrate values

FIBTGg_combined

Imaginary Fibre values

ALCg

Imaginary Alcohol values

ASHg

Imaginary Ash values

THIAmg

Imaginary Thiamine values

THIAHCLmg

Imaginary Thiamine from HCL values

RETOLmcg

Imaginary retinol values

CARTBEQmcg_combined

Imaginary Beta-Carotene equiv. values

NIAmg

Imaginary Niacin values

TRPmg

Imaginary Tryptophan values

NIAEQmg

Imaginary Niacin equivalent values

NIATRPmg

Imaginary Niacin from Tryptophan values

FATg

Imaginary Fat values

FAT_g

Imaginary Fat from unknown source values

FATCEg

Imaginary Fat from CE method values

comments

Imaginary comments

Source

An imaginary dataset; no formal source.